git-svn-id: https://svn.viathinksoft.com/svn/filter_foundry/trunk@573 1b5a0048-3058-4200-831f-25aa9e3a14b9

+++ ui_win.c
@@ -222,7 +222,7 @@ LRESULT CALLBACK ControlTextWndProc(HWND hWnd, UINT uMsg, WPARAM wParam, LPARAM 
 }
 
 #define SLIDER_EVENT(lParam) (((lParam) >> 16) & 0xFFFF)
-#define SLIDER_POS(lParam) ((lParam) & 0xFFFF)
+#define SLIDER_POS(lParam) ((int16_t)((lParam) & 0xFFFF))
 typedef enum PluginDllSliderEventType_ {
 	SliderEventMoved = 0x7FFB,
 	SliderEventReleased = 0x7FFC,
